Modify and rename: makefile and folders
The folders in data were modified to have the same name of the ones in pipeline. This makes the interpretation clear.
The python paths in the files were already fixed.
Moreover, the gathering tasks were added to the makefile.
The makefile cleaning is made by the command `make clean_{foldername}` and the build is done by the command `make {foldername}_`.